Remove retrying on supposed NTP errors
We only get an NTP timestamp once, at the beginning of the program.
It's more likely that this fails due to errors that the user introduces
(supplied disallowed port number, machine not connected to the network)
rather than due to NTP errors (where retries could actually help).

Thus, I choose to fail verbosely, and not spin in a futile retry loop.